refactor(build): split platform-specific build logic
This change extracts platform-specific Gradle configurations (Android, iOS, JVM) into separate files. This improves maintainability and readability of the build logic by keeping the main multiplatform script focused on shared configuration.
